# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is server of ctsv",
        "title": "CTSV API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/admin/student-address/update": {
            "put": {
                "description": "Admin update address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin with address"
                ],
                "parameters": [
                    {
                        "format": "model.Address",
                        "description": "List address need update",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Address"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/admin/student-class/create": {
            "post": {
                "description": "Admin create class",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin with class"
                ],
                "parameters": [
                    {
                        "format": "model.Class",
                        "description": "class",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Class"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/admin/student-class/update": {
            "put": {
                "description": "Admin update class",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin with class"
                ],
                "parameters": [
                    {
                        "format": "model.Class",
                        "description": "list class need update",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Class"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/admin/student-overview/update": {
            "put": {
                "description": "Admin update overview student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin with overview student"
                ],
                "parameters": [
                    {
                        "format": "model.OverView",
                        "description": "List Overview need update",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.OverView"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/admin/student-resultsOfAnnouncement/update": {
            "put": {
                "description": "Admin update ResultsOfAnnouncement",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin with ResultsOfAnnouncement"
                ],
                "parameters": [
                    {
                        "format": "model.ResultsOfAnnouncement",
                        "description": "list ResultsOfAnnouncement need update",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ResultsOfAnnouncement"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/admin/teacher-account/create": {
            "post": {
                "description": "Admin create account teacher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin with account teacher"
                ],
                "parameters": [
                    {
                        "format": "model.Teacher",
                        "description": "List account teacher",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Teacher"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "model.Address": {
            "type": "object",
            "properties": {
                "addressContact": {
                    "type": "string"
                },
                "baseInfoId": {
                    "type": "integer"
                },
                "currentResidentialAddress": {
                    "type": "string"
                },
                "permanentAddress": {
                    "type": "string"
                },
                "placeOfBirth": {
                    "type": "string"
                },
                "studentBaseInfo": {
                    "$ref": "#/definitions/model.StudentBaseInfo"
                }
            }
        },
...
        "model.Class": {
            "type": "object",
            "properties": {
                "classCode": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.OverView": {
            "type": "object",
            "properties": {
                "baseInfoId": {
                    "type": "integer"
                },
                "bloodGroup": {
                    "type": "string"
                },
                "cccdNumber": {
                    "type": "string"
                },
                "dateOfAdmissionToTheUnion": {
                    "type": "string"
                },
                "dateRange": {
                    "type": "string"
                },
                "decisionDay": {
                    "type": "string"
                },
                "ethnic": {
                    "type": "string"
                },
                "healthInsuranceCardNumber": {
                    "type": "string"
                },
                "issuedBy": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "numberOfDecisionsToJoinTheParty": {
                    "type": "string"
                },
                "partyAdmissionDate": {
                    "type": "string"
                },
                "policyObject": {
                    "type": "string"
                },
                "priorityObject": {
                    "type": "string"
                },
                "religion": {
                    "type": "string"
                },
                "studentBaseInfo": {
                    "$ref": "#/definitions/model.StudentBaseInfo"
                }
            }
        },
...
        "model.ResultsOfAnnouncement": {
            "type": "object",
            "properties": {
                "anh": {
                    "type": "number"
                },
                "award": {
                    "type": "string"
                },
                "baseInfoId": {
                    "type": "integer"
                },
                "combination": {
                    "type": "string"
                },
                "congDan": {
                    "type": "number"
                },
                "dia": {
                    "type": "number"
                },
                "form": {
                    "type": "string"
                },
                "hoa": {
                    "type": "number"
                },
                "industryCode": {
                    "type": "string"
                },
                "industryName": {
                    "type": "string"
                },
                "internationalCertificate": {
                    "type": "string"
                },
                "li": {
                    "type": "number"
                },
                "sinh": {
                    "type": "number"
                },
                "studentBaseInfo": {
                    "$ref": "#/definitions/model.StudentBaseInfo"
                },
                "su": {
                    "type": "number"
                },
                "toan": {
                    "type": "number"
                },
                "totalScoreForAdmission": {
                    "type": "string"
                },
                "van": {
                    "type": "number"
                }
            }
        },
...
        "model.Student": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pass": {
                    "type": "string"
                }
            }
        },
        "model.StudentBaseInfo": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "class": {
                    "type": "string"
                },
                "course": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "educationProgram": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "gender": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "student": {
                    "$ref": "#/definitions/model.Student"
                },
                "studentCode": {
                    "type": "string"
                },
                "studentId": {
                    "type": "integer"
                },
                "teacher": {
                    "$ref": "#/definitions/model.Teacher"
                },
                "teacherId": {
                    "type": "integer"
                },
                "trainingSystem": {
                    "type": "string"
                }
            }
        },
...
        "model.Teacher": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pass": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    }
}